requirements.txt
tensorflow==2.9.0
keras==2.9.0
numpy
pandas
matplotlib
opencv-python
Pillow
scikit-image
scipy
keras-retinanet
segmentation-models
plotly
ipykernel
python-dotenv
